+++ device-api/CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 ### Added
+- [SW-16508] Adding P2P DMA read/write commands and responses
 ### Changed
 ### Deprecated
 ### Removed